backend/requirements.txt

fastapi
uvicorn
langchain
python-dotenv==1.0.0
flask_cors
xmltodict
pinecone-client==3.1.0
google-generativeai
cohere
angle-emb
psycopg2-binary

